+++ R/convert_to_na.R
@@ -9,6 +9,8 @@
 #'   vectors or logicals) with values that should be converted to `NA`. Numeric
 #'   values applied to numeric vectors, character values are used for factors,
 #'   character vectors or date variables, and logical values for logical vectors.
+#' @param drop_levels Logical, for factors, when specific levels are replaced
+#'   by `NA`, should unused levels be dropped?
 #' @param ... Not used.
 #' @inheritParams find_columns
 #'
@@ -74,8 +76,9 @@ convert_to_na.numeric <- function(x, na = NULL, verbose = TRUE, ...) {
 }
 
 
+#' @rdname convert_to_na
 #' @export
-convert_to_na.factor <- function(x, na = NULL, verbose = TRUE, ...) {
+convert_to_na.factor <- function(x, na = NULL, drop_levels = FALSE, verbose = TRUE, ...) {
   # if we have a list, use first valid element
   if (is.list(na)) {
     na <- unlist(na[sapply(na, is.character)])
@@ -90,6 +93,9 @@ convert_to_na.factor <- function(x, na = NULL, verbose = TRUE, ...) {
     x[matches] <- NA
     # drop unused labels
     value_labels <- attr(x, "labels", exact = TRUE)
+    if (is.factor(x) && isTRUE(drop_levels)) {
+      x <- droplevels(x)
+    }
     attr(x, "labels") <- value_labels[!value_labels %in% na]
   }
   x
@@ -142,10 +148,10 @@ convert_to_na.logical <- function(x, na = NULL, verbose = TRUE, ...) {
 
 #' @rdname convert_to_na
 #' @export
-convert_to_na.data.frame <- function(x, na = NULL, select = NULL, exclude = NULL, ignore_case = FALSE, verbose = TRUE, ...) {
+convert_to_na.data.frame <- function(x, na = NULL, drop_levels = FALSE, select = NULL, exclude = NULL, ignore_case = FALSE, verbose = TRUE, ...) {
   # evaluate arguments
   select <- .select_nse(select, x, exclude, ignore_case, verbose = verbose)
 
-  x[select] <- lapply(x[select], convert_to_na, na = na, verbose = verbose, ...)
+  x[select] <- lapply(x[select], convert_to_na, na = na, drop_levels = drop_levels, verbose = verbose, ...)
   x
 }