+++ tests/testthat/test-convert_to_na.R
@@ -8,6 +8,14 @@ test_that("convert_to_na-factor", {
   x <- convert_to_na(iris$Species, na = list(2, "versicolor"))
   expect_equal(sum(is.na(x)), 50)
 
+  x <- convert_to_na(iris$Species, na = list(2, "versicolor"), drop_levels = FALSE)
+  expect_equal(levels(x), c("setosa", "versicolor", "virginica"))
+  expect_equal(as.vector(table(x)), c(50, 0, 50))
+
+  x <- convert_to_na(iris$Species, na = list(2, "versicolor"), drop_levels = TRUE)
+  expect_equal(levels(x), c("setosa", "virginica"))
+  expect_equal(as.vector(table(x)), c(50, 50))
+
   x <- suppressWarnings(convert_to_na(iris$Species, na = 2))
   expect_warning(convert_to_na(iris$Species, na = 2))
   expect_equal(sum(is.na(x)), 0)
@@ -58,6 +66,10 @@ test_that("convert_to_na other classes", {
   x <- convert_to_na(d$b, na = "c")
   expect_equal(x, structure(c(1L, 2L, NA, 4L, 5L),
                             .Label = c("a", "b", "c", "d", "e"),
+                            class = "factor"), tolerance = 1e-3, ignore_attr = TRUE)
+  x <- convert_to_na(d$b, na = "c", drop_levels = TRUE)
+  expect_equal(x, structure(c(1L, 2L, NA, 3L, 4L),
+                            .Label = c("a", "b", "d", "e"),
                             class = "factor"), tolerance = 1e-3, ignore_attr = TRUE)
   x <- convert_to_na(d$c, na = "2022-03-22")
   expect_equal(x, structure(c(NA, 18994, 19025, 18719, 18280), class = "Date"), tolerance = 1e-3, ignore_attr = TRUE)